This allows starting Tokio runtime only when necessary. 
Mean execution time for the same input drops from 149.3ms to 61.3ms:

Before:

```shell
> hyperfine --warmup 20 "git diff --patch | ./target/release/blockwatch"
Benchmark 1: git diff --patch | ./target/release/blockwatch
  Time (mean ± σ):     149.3 ms ±  12.3 ms    [User: 112.3 ms, System: 15.8 ms]
  Range (min … max):   138.9 ms … 192.2 ms    19 runs
```

After:

```shell
> hyperfine --warmup 20 "git diff --patch | ./target/release/blockwatch"
Benchmark 1: git diff --patch | ./target/release/blockwatch
  Time (mean ± σ):      61.3 ms ±   0.7 ms    [User: 58.3 ms, System: 10.4 ms]
  Range (min … max):    59.8 ms …  63.0 ms    46 runs
```